Change in picking default tenant for a user
Default tenant (if not provided explicitly through the http API) will now be taken from user, instead of directly from case system.
If user belongs to only a single tenant, then that tenant will be used as default for that user, otherwise, if user belongs to no tenant or to multiple tenants, then default system tenant as configured will be taken. This may lead to subsequent errors if user is not member of that tenant (which was already the case in previous behavior)