Distinguish what must run first and what must run later
PlanItemTransitions may impact interested sentries and stage completion check, leading to new events.
The execution of the behavior of these events may be recursive.
In this commit, the execution of entry criteria and stage-completion plus exit criteria is split up into 2 phases.
Entry criteria are investigated as soon as the PlanItemTransition happens.
Stage completion and exit criteria only after all potential follow-up entry criteria have also been executed, thereby
guaranteeing that subsequently activated plan items actually get created, such that stage completion does not ignore those.

This commit effectively makes the order of plan items in the XML definition of a stage no longer relevant.
That in itself will make the expected behavior of the engine more intuitive to developers using the engine.
To be more concrete: in the IDE, the order in which the various elements are drawn onto the screen no longer effects the actual execution.