Extended XMLHelper to trim whitespace around CDataSections
Expressions in the XML definition of the case model are put inside CDataSections.
When these are serialized in an element by the IDE, these sections get newline TextNodes around them in the parent element.
When using default org.w3c.dom parentElement.textContent() method, this leads to unnecessary and confusing extra newlines (as these are not visible in the IDE).
This fix ignores those newlines.