Removed "behavior" from ProcessTask events
moved "behavior" back into processtask actor itself;
note that the behavior (mostly invoking SubProcess implementation) still requires the actual state to be updated first. I believe it would be better if the event is generated only after that call is made, but that is not so much the purpose of this commit.

With this commit, EventCallStack behavior has been limited to CaseFileEvent and PlanItemTransitioned event, i.e., it has now been reduced to only the complex sentry network evaluation stack.