Refactor CaseTransaction code
Handling of PlanItemCreated code was duplicated.
Also harmonized the event handling pattern, now it is only method calls.
Also case instance now is only fetched when needed (which is still always, but at least it is now more to the point: fetched when needed instead of always at the beginning)