Case Commands now also leverage CaseLastModified
When invoking StartCase with a team, and directly sending a command to change the team, then the query db may not yet be up-to-date to for tenant access authorization.
This is now fixed by also checking on optional value of Case-Last-Modified header.

Also renamed the internal getCaseAndTenantInformation into authorizeTaskAccess and authorizeCaseAccess, as that is actually what is happening inside the implementations