Avoid NullPointer on persistence failures
Removed "currentCommand" member as it is never filled with sensible data.
Also removed assumption that it always has a value; replaced this with a check on whether there is a ModelCommand available,
and if so, use that to tell the sender about the issue. Otherwise leave it to basic error logging.
Also made the sender() information localized to the handling of the current message by introducing a recipient on ModelResponse
that is set with a value during creation of the ModelResponse (to keep avoid potential async "overwriting" of global Actor value)

Fixes #123