Property validation for task output binding to case file
With the case-file-enhancements commit, task output validation was removed.
This makes it possible to store "undefined" data in the case file, which is sometimes useful, e.g. when storing and handling ProcessTask errors.
However, the data that _is_ defined, must be validated. With this new commit we now validate properties against their defined types.

So, it is allowed to pass non-existing properties, but properties with a definition are validated for their type.