Add option to show debug information upon invalid commands
If commands are parsed and created, sometimes it is already clear that the command is invalid, and it immediately raises the exception.
This exception is catched by the AuthenticatedRoute and the message is sent back to the client. With this new option, when debug logging is enabled in the engine, the stacktrace is also printed in the logfile.