New event for RemoveChild during ReplaceCaseFileItem
When a CaseFileItem is replaced, those elements that are not in the new data but where available in the existing item now get a removeChild transition and event.
Includes refactoring of case file change propagation. No longer through the methods 'propagateValueChangeToChildren' and 'adoptContentFromParent', as they were sort of
indirectly assuming some context, whereas the initating logic has a much better view of the context and what to do.

Also refactored CaseFileMerger in the projection to become a little smarter
-- only handle events at top level of paths (optimization)
-- resolving logic and path parsing can now be done inside Path element, StringPath
 is no longer needed (it was duplicate code)