Fix invalid tenant recovery error
If e.g. a Tenant is recovered (but one that was never created) then ModelActor tries to set the tenant even for a DebugEvent, causing a next command to that Tenant to throw a different kind of error.

This code can be tested by running two times below (with interval at least 5 seconds)
"npm run dev 100 testtenantregistration"
Note: make sure that the ModelActor is removed from memory during the interval. E.g. by setting cafienne.actor.idle-period to 5s in the developer.conf