Enable "empty" output implementation parameters for expressions
In case of parameter mapping expressions that do not need an actual raw output parameter (e.g., "task.id" or "user.name"), there is no need to provide that parameter.
If the case definition leaves the sourceRef attribute empty on that element, it will check whether a non-empty transformation element is available.
If that is the case, the transformation will be executed with a Value.NULL as the parameter value, and the expression can return it's outcome.