Fix reading case file item children in input mappings
CaseFileItem children are not created before the parent is created, hence if they are used in input mappings upon creation of a parent object, they will be valued as null (even though the parent item has the proper value). This is because ValueMap tries to be too "intelligent"... Simply getting fixes the issue.
Fixes #185