Fix recovery issue for CaseFile
Massive issue, actually, oops :(
Upon recovery resolving existing array items was handled "nicely" in CaseFileItemArray, but in or around commit 9309b1f825f15b4db74a0c95f18bbe6d161d0649 this logic seemed to sort of get skipped. And it was not detected. We need to do compatibliyt testing on recovery of cases as well!!!