Support execution within Case instead of ProcessTaskActor
Made a new InlineSubProcess class and Definition for it, that can be used to run a process task inline, i.e. without creating a new akka persisten actor instance for it.
Note, this requires the process to be executed on the current thread or for it to handle it's own execution context (not at all recommended, NOT at all).
Calculation Task is actually only doing some SPEL expressions and therefore need not have it's own independent running. It may also help scenarios that returning case last modified actually also incorporated the result of the calc task.