Restructure CasePlan to be less definition dependent
Basic approach is to change instance constructors, to no longer set final fields based on definition.
Instead add a method to retrieve the value from the definition when it is required.
That way, the values are always up-to-date with the latest definition.

Also move CasePlan creation back inside Case again.