Block actions on child CaseFileItems if parent is not in state Available
When the parent is not created, or the parent is deleted, it should not be possible to do actions on the child.
Earlier this was possible, e.g. create a child and after that create the parent, but the projection fails on that situation (parent does not take child value into account and drops it).

Note: this also requires some changes in the internal unit tests