Fix 'user' access in spel expressions, preparing for consent groups
User no longer exposes name, email and roles. These need to be taken from the case file context.
For a case, the user property now gives his membership information, i.e., the case roles.
Earlier the 'user' property was available one very APIObject, but that is not logical. Now it is only available on the root level.
This avoids conflicts with e.g. a case file item that has a property named 'user'.