Enable platform access for all users with a valid token
Instead of 401 Unauthorized, these users will now get mostly 404 Not Found.
Note: obviously these users are able to get a valid token to access the case engine, so there is no specific need to raise Unauthorized, Not Found is good enough.

On the longer term this enables inviting non-platform registered users to a case, in a manner similar to anonymous start case support.