Refactor CaseSystem.router() to CaseSystem.gateway
For sending messages into the case system, the method CaseSystem.router returned an ActorRef (enabling e.g. to run the akka ask pattern).
This is now replaced with a wrapper class CafienneGateway, so that the code for sending and asking commands to the case system is in one place.
Also removed the not-working cluster router.