Remove deprecated code no longer in use
Remove snapshot based storage recovery support for timer service (it only confuses the code and is only in use up to version 1.1.12). People still on that version can contact us and we'll happily support them to 1.1.18.
Also renamed "Scheduled.scala" to "TimerJob", as that name is now free, and it also gives a slightly better description of what is done.