Fix description generation
This code sets the gitDescribedVersion from the current git tag.
We do not actually return a version, but the version + the description.
Next task is to pull them apart from each other again.
This is done because somehow it is not possible to keep state inside the GitInfoAnalyzer object :(

Also removed a compilation warning