Move to batched-wise ModelEvents to fill QueryDB
Sinks that fill the QueryDB (such as CaseEventSink, TenantEventSink) now receive a batch instead of individual ModelEvents.
This enables sharper transaction handling.
Note, a batch is the resulting set of events that a ModelActor stores upon handling a ModelCommand.
It always "completes" with a CommitEvent.

TaggedEventConsumer is converted to TaggedEventSource, basically providing a Source[ModelEvent, NotUsed].
TaggedEventBatchSink extends the TaggedEventSource.

TimerService does not need batches, just the individual events. Therefore it uses TaggedEventSource directly,
and some of the logic of former TaggedEventConsumer is moved in here.

Also renamed SlickEventMaterializer to QueryDBEventSink, as that matches better with CaseEventSink, TenantEventSink, etc.