Make Storage again an input to the QueryDB Sinks
Added a new concept 'Storage' that can map Records to Slick QueryTables.
E.g., CaseStorage can store records that need to go to the Query DB when Case Events are handled by the CaseEventSink.

Note, the Storage does not store the records itself, it can provide "db transactions" responsible for the actual storage.

This layer enables re-enabling the various scala unit tests on the actual logic inside the sinks (e.g., which records are created from the events that result from starting a case).

The generic "Persistence" trait that existed earlier is now replaced with sink-specific traits. E.g., CaseEventSink needs a CaseStorage, and TenantEventSink needs a TenantStorage.

This is a step to make the records explicit. The earlier "Persistence" trait had an upsert(record: Any) method, which is too generic.