Add new events to track inter-model communication
When a case invokes a sub process or sub case, failures were only sent to the log file. With this commit this now generates an event on the case, with the status of the communication. This also helps in understanding what the status of the "other" ModelActor is.
Note, also removed askCase and askProcess and replaced it with informImplementation and informParent, as that is a more logical description of the underlying use cases.
Note: events for failing communication initiated from the task implementation are to be added.