Create new route + port old routes to it
PlanItemHistory is now created dynamically based on the latest & greatest case events, instead of going through database queries. Also deleted the database queries. Tables are yet to be done. Also records and case classes and merger can all be brought together to a single place (requiring some internal refactoring).
This step is another step towards simplifying the CasePlanEvent json structure.

This commit also creates a new trait InstanceEventSource, which is sort of an alternative to TaggedEventSource.