Reset git local changes publishing docker image
The local changes are disturbing the build creation upon release, and seem also not to be necessary at all.
Disturbing --> when release build is done, it is multi-architecture, which requires creating the image again (sbt docker / stage).
At that moment, sbt detects local git changes and uses that to modify the Cafienne Engine version, which is not supposed to happen.