More controlled stage activation
Fixes #365
When a stage becomes Active, it instantiates it's children.
If those children included Milestones that immediately Occur, this may trigger behavior on subsequent children that are created later on.
Therefore we now first create the java instance of the PlanItem, and in a second step add it to the sentry network and trigger the Create transition.

Note: we also moved the code for creating repeat items from PlanItem to Stage, to have more control on when to trigger the second step; "more control" means: only within the stage and not somewhere else.