Use BoardQueries instead of TaskQueries
TaskQueries.getAllTasks actually filters on user's roles, and to get the full board, also tasks that cannot be performed by the user still must be returned and visualized.
Note that for case instance this already works, but for board flows it does not work, unless we'd do a query for each flow in the board. Therefore adding a query that can retrieve all tasks in all flows in the board.
Note: we could consider moving something similar into TaskQueries as an addition to GetCaseTasks (e.g. getAllTasksInCasesWithMembership or so)