More robust handling of lists with null elements
Deserializing a command that has a set of String (typically roles of user)  with one or more null elements leads to a DeserializationFailure due to the implementation in ValueMap's readStringList method.
Added and used explicit ApiValidator for this.
DeserializationFailures are now also captured and logged more explicitly inside the CaseMessageRouter.