Small refactoring on Stage completion check
Generalized the handling of a transition on a plan item in its parent stage.
The PlanItem now just informs the parent Stage about its event, and leaves all completion logic to the stage.
With that, a Stage can now also try to complete itself without generating a dummy event

Also:
- Debug logging when adding a planitem to the sentry network
- Better debug logging about the reason when the Stage triggers 'create' transition on children
- Handy comparison methods on the Transition enum
- Some updates to clean up code and javadoc