Give explicit feedback on running storage process
Query DB is not always immediately up-to-date (with e.g. a deleted case in a tenant).
The event journal may already be having an in-progress StorageEvent that blocks recovery of the original ModelActor. This is now presented in a functional manner to the sender of the command.
The CommandRoute then returns a search failure similar to what the QueryRoute does.
Note that for sending bootstrap commands to such an actor the logic remains as is (i.e., resulting in 400 BadRequest)
We now also log additional info on the user and the message that caused the recovery failure