Enhance case team migration
When updating the definition of a case, the case first updates the team definition, then the file definition, then the plan definition.
A new case team can be passed, and until now that new team was only set after the plan was updated, resulting in strange behavior. Also that new team was not yet propagated to subcases.

The new team is now set while migrating the case team definition. So it is immediately active also for e.g. assignment of new human tasks that come from the new definition.

This fixes #390