Recover state of sentry network
Earlier, the sentry network would not run logic on updates of events during recovery of the case.
This leads to issue #395.
This commit fixes #395.

From now on, the sentry network propagates the state changes to the elements inside, and the individual elements are responsible for updating the right state during recovery.