Enhanced CMMN Fault Handling
CMMN has very limited and even error-prone fault handling on e.g. process task failures. If this is not explicitly modeled, the case sort of ignores the failure, as Failed is a semi-terminal state.
Most important background reason for this, is that "stage completion is tried whenever a child item transitions to semi-terminal state."
In practice this leads too often to totally unexpected fault handling behavior for people that model cases. In addition, modeling explicit fault handling is not straightforward.

With this commit, the Cafienne engine enhances the default CMMN behavior with more intuitive fault handling. Through an engine wide config setting it is possible to support the existing CMMN behavior as before.
When a plan item goes into Failed through the "fault" transition, then the stage surrounding it _also_ goes into Failed state (if it was not in that state already).
When a plan item is reactivated, then also this transition is propagated to the stage surrounding it, and if none of the other items in that stage are in Failed state, the stage will be reactivated as well (and then becomes Active again).
When a stage is reactivated, this is propagated into children that are in Failed state. This is existing behavior, but through the bubbling it now becomes possible to reactivate all failing elements in a case in one shot, by sending the reactivate to the case itself.
This behavior even propagates into the chain of cases and subcases.

Note: the bubbling of the transition is done in the delayed behavior of the event, but still before the item's exit criteria are informed. Note also that stage completion is no longer tried if the fault is bubbled into the stage.
As before this is only done when not recovering.