Cleanup groups during tenant deletion - if they are not in use
During deletion of a tenant, it removes events from the journal from all cases and groups inside the tenant.
Also, case and tenant data is removed from the query database.
This commit now also removes the consent group information from the query database.

Furthermore an extra validation is done on the tenant that it cannot be deleted if one of its consent groups is still in use in the team of a case in a different tenant.

Fixes #406
Fixes #407