Improve PlanItemOnPart migration
If a case is in migration, the connection of criteria to other plan items no longer compares on equality of definition, but on similarity.
So, if a connection to another plan item is to a plan item that is already (or not yet) migrated, then now the connection will still be done.