Remove duplicate code
The layer between Queries and Routes has some duplicate case classes here and there.
This was originally intended to use this to create Swagger descriptions, but this has been replaced with the example formats in many places.
Therefore removing some of the duplicate code in this commit.

Also removed the XML extension in the base QueryRoute, as it is in use only in 1 place.