Extend core UserIdentity to have additional roles
This helps in defining mechanisms to "bypass" TenantUser registration.
Queries that read tenant roles on user level require such registration, but they are now also able to read the roles from the UserIdentity instead.
In order to use this, one must extend the UserIdentity trait when executing queries or commands. This functionality is not exposed through the web api