Also implicitly create array items for cfi onParts
When an entry criterion (or any other criterion) has an onPart that listens to a case file item, it will try to connect to that item.
The current onPart logic assumes that items will be created implicitly for such scenarios.
This logic fails for array type of case file items and specifically their children.
With this fix that is done anyway.

Fixes #478