Extend task queries to select executable tasks
With this new flag it is possible to filter task queries to retrieve only tasks that can be performed by the user, instead of e.g. all case tasks.
For GetCaseTasks we also return a new boolean field indicating whether the users requesting the tasks is allowed to perform them or not

Also refactored the file structure a bit, to have more clear separation of logic